+++ lib/rubocop/ast/node/str_node.rb
@@ -6,6 +6,8 @@ module RuboCop
     # in place of a plain node when the builder constructs the AST, making
     # its methods available to all `str` nodes within RuboCop.
     class StrNode < Node
+      include BasicLiteralNode
+
       def heredoc?
         loc.is_a?(Parser::Source::Map::Heredoc)
       end